requirements.txt
httpx>=0.24.0
mcp>=1.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0 